+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Fetch states for selected country deployments
 - [In Progress: 31 Aug 2021] Record call results instead of updating the same appointment record
 - [In Progress: 31 Aug 2021] Refactor logic around providing drug frequencies label depending on the country
+- Remove login failed errors from UiRenderer, and add via effect handler when logging in
 
 ### Changes
 - Implement providing drug frequencies label depending on the country